requirements.txt
asgiref==3.2.3
Django==3.0.1
djangorestframework==3.11.0
numpy==1.18.0
pandas==0.25.3
psycopg2==2.8.4
python-dateutil==2.8.1
pytz==2019.3
six==1.13.0
SQLAlchemy==1.3.12
sqlparse==0.3.0
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
certifi==2019.9.11
chardet==3.0.4
idna==2.8
importlib-metadata==1.1.0
more-itertools==8.0.0
multidict==4.6.1
packaging==19.2
pluggy==0.13.1
py==1.8.0
pyparsing==2.4.5
pytest==5.3.1
requests==2.22.0
six==1.13.0
urllib3==1.25.7
wcwidth==0.1.7
yarl==1.3.0
zipp==0.6.0